Split helm package into several files
